# openapi.yaml
openapi: 3.0.3
info:
  title: EBKO-CRM API
  version: 1.0.0
  description: REST API для системы управления обращениями EBKO-CRM
servers:
  - url: https://vk-dev.fun
    description: Production сервер
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Числовой идентификатор ресурса
  responses:
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "employee not found"

    # ── Auth ─────────────────────────────────────────────────────────────────
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          example: admin
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: "a3f8c1d2e4b5..."
        employee:
          $ref: '#/components/schemas/Employee'

    # ── Employee ──────────────────────────────────────────────────────────────
    Employee:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "ivan.ivanov"
        full_name:
          type: string
          example: "Иванов Иван Иванович"
        email:
          type: string
          nullable: true
          example: "ivan@company.ru"
        phone:
          type: string
          nullable: true
          example: "+7 900 000-00-00"
        role:
          type: string
          example: "manager"
        photo_url:
          type: string
          nullable: true
          example: "/uploads/employee-1-1234567890.jpg"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EmployeeCreate:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: "ivan.ivanov"
        password:
          type: string
          example: "secret123"
        full_name:
          type: string
          example: "Иванов Иван Иванович"
        email:
          type: string
          example: "ivan@company.ru"
        phone:
          type: string
          example: "+7 900 000-00-00"
        role:
          type: string
          example: "manager"
...
    Client:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "ООО Ромашка"
        inn:
          type: string
          nullable: true
          example: "7701234567"
        email:
          type: string
          nullable: true
          example: "info@romashka.ru"
        phone:
          type: string
          nullable: true
          example: "+7 495 000-00-00"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ClientCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "ООО Ромашка"
        inn:
          type: string
          example: "7701234567"
        email:
          type: string
          example: "info@romashka.ru"
        phone:
          type: string
          example: "+7 495 000-00-00"
...
    Site:
      type: object
      properties:
        id:
          type: integer
          example: 1
        client_id:
          type: integer
          example: 1
        product_id:
          type: integer
          nullable: true
          example: 2
        name:
          type: string
          example: "Офис на Ленина"
        address:
          type: string
          nullable: true
          example: "г. Москва, ул. Ленина, д. 10"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SiteCreate:
      type: object
      required: [client_id, name]
      properties:
        client_id:
          type: integer
          example: 1
        product_id:
          type: integer
          nullable: true
          example: 2
        name:
          type: string
          example: "Офис на Ленина"
        address:
          type: string
          example: "г. Москва, ул. Ленина, д. 10"
...
    Equipment:
      type: object
      properties:
        id:
          type: integer
          example: 1
        product_id:
          type: integer
          nullable: true
          example: 1
        name:
          type: string
          example: "Домофон Vizit"
        model:
          type: string
          nullable: true
          example: "БВД-405"
        description:
          type: string
          nullable: true
          example: "Видеодомофон с памятью"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EquipmentCreate:
      type: object
      required: [name]
      properties:
        product_id:
          type: integer
          nullable: true
          example: 1
        name:
          type: string
          example: "Домофон Vizit"
        model:
          type: string
          example: "БВД-405"
        description:
          type: string
          example: "Видеодомофон с памятью"
...
    Ticket:
      type: object
      properties:
        id:
          type: integer
          example: 1
        type:
          type: string
          enum: [request, complaint, incident, task]
          example: "complaint"
        status:
          type: string
          enum: [open, in_progress, pending, resolved, closed]
          example: "open"
        priority:
          type: string
          enum: [low, medium, high, critical]
          example: "high"
        product_id:
          type: integer
          nullable: true
          example: 1
        description:
          type: string
          example: "Не работает домофон в подъезде №3"
        client_id:
          type: integer
          nullable: true
          example: 1
        site_id:
          type: integer
          nullable: true
          example: 2
        responsible_id:
          type: integer
          nullable: true
          example: 3
        deadline:
          type: string
          format: date-time
          nullable: true
          example: "2026-03-10T18:00:00Z"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TicketCreate:
      type: object
      properties:
        type:
          type: string
          enum: [request, complaint, incident, task]
          default: request
        status:
          type: string
          enum: [open, in_progress, pending, resolved, closed]
          default: open
        priority:
          type: string
          enum: [low, medium, high, critical]
          default: medium
        product_id:
          type: integer
          nullable: true
        description:
          type: string
          example: "Не работает домофон в подъезде №3"
        client_id:
          type: integer
          nullable: true
        site_id:
          type: integer
          nullable: true
        responsible_id:
          type: integer
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
...
    Dashboard:
      type: object
      properties:
        id:
          type: integer
          example: 1
        employee_id:
          type: integer
          example: 2
        name:
          type: string
          example: "Мои открытые обращения"
        filters:
          type: object
          example: { "status": "open", "responsible_id": 2 }
        position:
          type: integer
          example: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DashboardCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Мои открытые обращения"
        filters:
          type: object
          example: { "status": "open", "responsible_id": 2 }
    DashboardUpdate:
      type: object
      properties:
        name:
          type: string
        filters:
          type: object
...
paths:
...
  /api/auth/login:
    post:
      tags: [Авторизация]
      summary: Вход в систему
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/logout:
    post:
      tags: [Авторизация]
      summary: Выход из системы
      responses:
        '200':
          description: Успешный выход
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "logged out"
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Tickets ────────────────────────────────────────────────────────────────
  /api/tickets:
    get:
      tags: [Обращения]
      summary: Список обращений
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, pending, resolved, closed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: type
          in: query
          schema:
            type: string
            enum: [request, complaint, incident, task]
        - name: client_id
          in: query
          schema:
            type: integer
        - name: responsible_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Список обращений
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Обращения]
      summary: Создать обращение
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreate'
      responses:
        '201':
          description: Обращение создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/employees:
    get:
      tags: [Сотрудники]
      summary: Список сотрудников
      responses:
        '200':
          description: Список сотрудников
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Сотрудники]
      summary: Создать сотрудника
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreate'
      responses:
        '201':
          description: Сотрудник создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Пользователь с таким username уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/clients:
    get:
      tags: [Клиенты]
      summary: Список клиентов
      responses:
        '200':
          description: Список клиентов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Клиенты]
      summary: Создать клиента
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreate'
      responses:
        '201':
          description: Клиент создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/sites:
    get:
      tags: [Площадки]
      summary: Список площадок
      parameters:
        - name: client_id
          in: query
          schema:
            type: integer
        - name: product_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Список площадок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Site'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Площадки]
      summary: Создать площадку
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteCreate'
      responses:
        '201':
          description: Площадка создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/equipment:
    get:
      tags: [Оборудование]
      summary: Список единиц оборудования
      parameters:
        - name: product_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Список оборудования
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Equipment'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Оборудование]
      summary: Создать единицу оборудования
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentCreate'
      responses:
        '201':
          description: Оборудование создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/dashboards:
    get:
      tags: [Дашборды]
      summary: Список дашбордов текущего пользователя
      responses:
        '200':
          description: Список дашбордов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dashboard'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Дашборды]
      summary: Создать дашборд
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardCreate'
      responses:
        '201':
          description: Дашборд создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/dashboards/{id}:
    parameters:
      - $ref: '#/components/parameters/id'

    get:
      tags: [Дашборды]
      summary: Просмотр дашборда
      responses:
        '200':
          description: Дашборд
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Дашборды]
      summary: Редактировать дашборд (название, фильтры)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardUpdate'
      responses:
        '200':
          description: Обновлённый дашборд
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Дашборды]
      summary: Удалить дашборд
      responses:
        '204':
          description: Дашборд удалён
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Profile ────────────────────────────────────────────────────────────────
  /api/profile:
    get:
      tags: [Профиль]
      summary: Просмотр своего профиля
      responses:
        '200':
          description: Профиль текущего пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
...